+++ src/Retlang/Core/IScheduler.cs
@@ -13,7 +13,7 @@ namespace Retlang.Core
         /// <param name="action"></param>
         /// <param name="firstInMs"></param>
         /// <returns>a handle to cancel the timer.</returns>
-        IDisposable Schedule(Action action, long firstInMs);
+        IDisposable Schedule(Action action, int firstInMs);
 
         /// <summary>
         /// Schedule an action to be executed on a recurring interval.
@@ -22,6 +22,6 @@ namespace Retlang.Core
         /// <param name="firstInMs"></param>
         /// <param name="regularInMs"></param>
         /// <returns>a handle to cancel the timer.</returns>
-        IDisposable ScheduleOnInterval(Action action, long firstInMs, long regularInMs);
+        IDisposable ScheduleOnInterval(Action action, int firstInMs, int regularInMs);
     }
 }
